Actually close the server, sthread, and client when `Bye.` is sent by the client.
Prior to this commit, when the client sent `Bye.`, the SThread would
close immediately before passing the message to the server. This
prevented the server from sending `BYE.` (the client was expecting
`Bye.` so this wouldn't have worked anyway) back to confirm to the
client that it closed and that the client is free to close as well.